Fix notification queue
