+++ src/core/renderers/NotificationsRenderer.tsx
@@ -9,7 +9,7 @@ import { useAnimationAPI } from '../hooks/useAnimationAPI'
 export const NotificationsRenderer = () => {
   const { config, ...state } = useNotificationsStates()
 
-  const animationAPI = useAnimationAPI(config, state?.notificationEvent?.id)
+  const animationAPI = useAnimationAPI(config)
 
   useNotificationEventHandler({ ...state, ...animationAPI })
