+++ src/types/reducer.ts
@@ -1,0 +1,11 @@
+import type { EmitParam, ModifiedEmitParam } from '../core/services/types'
+
+export type Actions = Add | Remove | Modify | Pop
+export type Add = { type: 'add'; payload: EmitParam }
+export type Remove = {
+  type: 'remove'
+  payload: EmitParam['id']
+  onActiveNotification: (id: EmitParam['id']) => void
+}
+export type Modify = { type: 'modify'; payload: ModifiedEmitParam<unknown> }
+export type Pop = { type: 'pop'; payload?: EmitParam['id'] }
